Refactored functions on the POST Route to Middleware NEED: ERROR HANDLING UTILITIES
